requirements.txt
# --- Core Frameworks ---
langchain>=0.1.0
langchain-community
langchain-google-genai
langchain-chroma
langchain-huggingface

# --- AI & NLP ---
google-generativeai
sentence-transformers
transformers
torch
protobuf
openai
langchain-core

# --- Data & Vector DB ---
chromadb
pypdf
PyMuPDF
docarray
tiktoken

# --- UI & Environment ---
streamlit
python-dotenv
fastapi
uvicorn

# --- Utilities ---
pydantic
numpy
setuptools

# --- MLOps & Tracking ---
mlflow

ragas
datasets


langchain-classic
rank-bm25
fastapi
uvicorn
python-multipart